services/billingService/requirements.txt
Flask==2.3.3
stripe==7.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0
pydantic==2.4.2
pytest==7.4.2